requirements.txt
pandas
numpy
langchain-community
langchain
beautifulsoup4
langchain-core
streamlit
# Environment Variable Management
python-dotenv
# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
pypdf
pydantic
grandalf